separated unix process and process logic
