+++ pkg/client/proc.go
@@ -5,66 +5,59 @@ import (
 	"runtime"
 
 	"capnproto.org/go/capnp/v3"
-	"github.com/wetware/ww/pkg/cap/proc"
 	"github.com/wetware/ww/pkg/cap/proc/unix"
 )
 
-type UnixProc struct {
+type UnixExecutor struct {
 	unix.Client
 }
 
-func (p UnixProc) Exec(ctx context.Context, name string, args ...string) *Process {
-	return newProcess(p.Client.Exec(ctx, name, args...))
-}
-
-type Process struct {
-	client  proc.Process
-	release capnp.ReleaseFunc
-}
-
-func newProcess(p proc.Process, release capnp.ReleaseFunc) *Process {
-	proc := &Process{client: p, release: release}
+func (p UnixExecutor) Exec(ctx context.Context, name string, args ...string) *UnixProcess {
+	client, release := p.Client.Exec(ctx, name, args...)
+	proc := &UnixProcess{client: client, release: release}
 	runtime.SetFinalizer(proc, release)
 	return proc
 }
 
-func (c Process) Start(ctx context.Context) error {
+type UnixProcess struct {
+	client  unix.Process
+	release capnp.ReleaseFunc
+}
+
+func (c UnixProcess) Start(ctx context.Context) error {
 	return c.client.Start(ctx)
 }
 
-func (c Process) Wait(ctx context.Context) error {
+func (c UnixProcess) Wait(ctx context.Context) error {
 	err := c.client.Wait(ctx)
-	c.release() // TODO: is this a correct beehavior?
+	c.release() // TODO: is this a correct behavior?
 	runtime.SetFinalizer(c, nil)
 	return err
 }
 
-func (c *Process) StderrPipe(ctx context.Context) *ReadCloser {
+func (c *UnixProcess) StderrPipe(ctx context.Context) *ReadCloser {
 	stderr, release := c.client.StderrPipe(ctx)
 	rc := &ReadCloser{client: stderr}
-
 	runtime.SetFinalizer(rc, release)
 	return rc
 }
 
-func (c *Process) StdoutPipe(ctx context.Context) *ReadCloser {
+func (c *UnixProcess) StdoutPipe(ctx context.Context) *ReadCloser {
 	stdout, release := c.client.StdoutPipe(ctx)
 	rc := &ReadCloser{client: stdout, release: release}
-
 	runtime.SetFinalizer(rc, release)
 	return rc
 }
 
-func (c *Process) StdinPipe(ctx context.Context) *WriteCloser {
+func (c *UnixProcess) StdinPipe(ctx context.Context) *WriteCloser {
 	stdin, release := c.client.StdinPipe(ctx)
 	rc := &WriteCloser{client: stdin, release: release}
-
 	runtime.SetFinalizer(rc, release)
 	return rc
 }
 
 type ReadCloser struct {
-	client  proc.ReadCloser
+	client  unix.ReadCloser
 	release capnp.ReleaseFunc
 }
 
@@ -80,7 +73,7 @@ func (rc *ReadCloser) Close(ctx context.Context) (err error) {
 }
 
 type WriteCloser struct {
-	client  proc.WriteCloser
+	client  unix.WriteCloser
 	release capnp.ReleaseFunc
 }
